Fixed image response handling
